+++ ui/src/radial_widget.cpp
@@ -6,12 +6,12 @@ using namespace std;
 
 bool RadialWidget::hit_test(const int x, const int y) const
 {
-  Float2 dpos = Float2(x, y) - center;
+  Float2 dpos = Float2((float)x, (float)y) - center;
   float r = dpos.magnitude();
   dpos = dpos / r; //normalize
 
   float theta = atan2(dpos[1], dpos[0]);
-  if(theta < 0.0f) { theta = 2.0f * M_PI + theta; }
+  if(theta < 0.0f) { theta = 2.0f * (float)M_PI + theta; }
 
   //if(arc[0] < arc[1])
   {
